usize → i32

+++ sample-jna/src/lib.rs
@@ -4,21 +4,22 @@ pub extern  fn sample_add( i: i32) ->i32 {
 }
 
 #[no_mangle]
-pub extern fn sieve_liner(n: usize) -> usize{
+pub extern fn sieve_liner(n: i32) -> i32{
     let mut primes = vec![];
-    let mut d = vec![0usize; n + 1];
+    let mut d = vec![0i32; n as usize + 1];
+    
     for i in 2..n + 1 {
-        if d[i] == 0 {
+        if d[i as usize] == 0 {
             primes.push(i);
-            d[i] = i;
+            d[i as usize] = i;
         }
         for p in &primes {
             if p * i > n {
                 break;
             } 
-            d[*p * i] = *p;
+            d[(*p * i) as usize] = *p;
         }
     }
     
-    primes.len() 
+    primes.len() as i32
 }
